+++ Source/EventFlow.Elasticsearch.Tests/Properties/AssemblyInfo.cs
@@ -28,11 +28,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("EventFlow.ReadStores.Elasticsearch.Tests")]
+[assembly: AssemblyTitle("EventFlow.Elasticsearch.Tests")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("EventFlow.ReadStores.Elasticsearch.Tests")]
+[assembly: AssemblyProduct("EventFlow.Elasticsearch.Tests")]
 [assembly: AssemblyCopyright("Copyright ©  2015")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]